wip build
currently this is still an issue, it's not upgrading to the newer
version if there is an old one available, even if it doesn't satisfy the
version_check it seems.
```
LoadError: Will not force WinRPM to rebuild dependency "libgeos".
Please make any necessary adjustments manually (This might just be a
version upgrade)
while loading C:\Users\visser_mn\.julia\v0.6\LibGEOS\deps\build.jl, in
expression starting on line 40
```